refactor: split datamachine_log into write and management actions (#206)
