+++ inc/Engine/Actions/DataMachineActions.php
@@ -50,6 +50,7 @@ use DataMachine\Engine\Actions\Handlers\MarkItemProcessedHandler;
 use DataMachine\Engine\Actions\Handlers\FailJobHandler;
 use DataMachine\Engine\Actions\Handlers\JobCompleteHandler;
 use DataMachine\Engine\Actions\Handlers\LogHandler;
+use DataMachine\Engine\Actions\Handlers\LogManageHandler;
 
 /**
  * Register core Data Machine action hooks.
@@ -61,7 +62,8 @@ function datamachine_register_core_actions() {
 	add_action( 'datamachine_mark_item_processed', array( MarkItemProcessedHandler::class, 'handle' ), 10, 4 );
 	add_action( 'datamachine_fail_job', array( FailJobHandler::class, 'handle' ), 10, 3 );
 	add_action( 'datamachine_job_complete', array( JobCompleteHandler::class, 'handle' ), 10, 2 );
-	add_action( 'datamachine_log', array( LogHandler::class, 'handle' ), 10, 4 );
+	add_action( 'datamachine_log', array( LogHandler::class, 'handle' ), 10, 3 );
+	add_action( 'datamachine_log_manage', array( LogManageHandler::class, 'handle' ), 10, 4 );
 
 	\DataMachine\Engine\Actions\ImportExport::register();
 	datamachine_register_execution_engine();
